frontend/requirements.txt
Flask==1.0.2
google-cloud-datastore==1.7.0
google-cloud-pubsub==0.36.0
